# src/main/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Swagger Title",
    "description": "설명!!",
    "contact": {
      "name": "사용자명",
      "url": "url",
      "email": "ih.park@fusionsoft.co.kr"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "v2"
  },
  "externalDocs": {
    "description": "더 많은 문서 보기",
    "url": "https://swagger.io/docs/"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "로컬 서버"
    }
  ],
  "tags": [
    {
      "name": "User v1",
      "description": "회원 API v1"
    },
    {
      "name": "Memo",
      "description": "메모 관련 API"
    },
    {
      "name": "users v2",
      "description": "회원 API v2"
    }
  ],
  "paths": {
    "/api/memo/modifyMemo": {
      "put": {
        "tags": [
          "Memo"
        ],
        "summary": "메모 수정",
        "description": "메모 내용 수정 .",
        "operationId": "modifyMemo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공적으로 수정됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          }
        }
      }
    },
    "/memos": {
      "get": {
        "tags": [
          "Memo"
        ],
        "summary": "전체 메모 조회",
        "description": "저장된 모든 메모를 조회합니다.",
        "operationId": "getMemos",
        "responses": {
          "200": {
            "description": "성공적으로 조회됨",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemoDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemoDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Memo"
        ],
        "summary": "메모 생성",
        "description": "새로운 메모를 저장합니다.",
        "operationId": "createMemo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/memo/insertMemo": {
      "post": {
        "tags": [
          "Memo"
        ],
        "summary": "메모 생성",
        "description": "새로운 메모를 저장합니다.",
        "operationId": "insertMemo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공적으로 저장됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          }
        }
      }
    },
    "/v2/users": {
      "get": {
        "tags": [
          "users v2"
        ],
        "summary": "v2 - 전체 회원 조회",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users": {
      "get": {
        "tags": [
          "User v1"
        ],
        "summary": "전체 회원 조회",
        "operationId": "getUsers_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/memos/{id}": {
      "get": {
        "tags": [
          "Memo",
          "Memo API"
        ],
        "summary": "메모 단건 조회",
        "description": "메모 ID를 이용해 특정 메모를 조회합니다.",
        "operationId": "getMemoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "조회할 메모의 고유 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          },
          {
            "name": "includeDetails",
            "in": "query",
            "description": "상세 정보를 포함할지 여부",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "example": true
          }
        ],
        "responses": {
          "200": {
            "description": "성공적으로 메모를 조회했습니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          },
          "404": {
            "description": "메모를 찾을 수 없습니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류입니다.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/memo/getMemo": {
      "get": {
        "tags": [
          "Memo"
        ],
        "summary": "단일 메모 조회",
        "description": "단일 메모.",
        "operationId": "getMemo",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적으로 조회됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/memo/getMemoList": {
      "get": {
        "tags": [
          "Memo"
        ],
        "summary": "전체 메모 조회",
        "description": "저장된 모든 메모를 조회합니다.",
        "operationId": "getMemos_1",
        "responses": {
          "200": {
            "description": "성공적으로 조회됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                },
                "examples": {
                  "메모 리스트 예제": {
                    "description": "메모 리스트 예제",
                    "value": [
                      {
                        "id": 1,
                        "content": "Test Memo 1"
                      },
                      {
                        "id": 2,
                        "content": "Test Memo 2"
                      }
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/memo/deleteMemo": {
      "delete": {
        "tags": [
          "Memo"
        ],
        "summary": "메모 삭제",
        "description": "메모 삭제 .",
        "operationId": "deleteMemo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "성공적으로 삭제됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemoDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MemoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "메모 ID",
            "format": "int32",
            "example": 1,
            "default": 1
          },
          "content": {
            "type": "string",
            "description": "메모 내용",
            "example": "Memo Swagger Test!",
            "enum": [
              "메모1",
              "메모2"
            ]
          }
        },
        "description": "메모 정보를 담는 DTO"
      }
    }
  }
}